Output column numbers for parse errors in INI files
The %locations directive tells bison to generate a
struct for tracking line and column numbers (or any
custom location info) for each token. The appropriate
location is passed to the yyerror function when an
unexpected token is encountered.

The lexer is responsible for populating this struct
each time a token is emitted. In this case, we
replace the existing line number global with struct
updates, and calculate the new column offset in the
special <*> rule.

The default bison code initialises the struct to all
ones, so the resulting column number is 1-indexed.