First pass at adding column number tracking to main parser
This struggles because the location calculation is entirely
at the beginning of the token processing, so we haven't yet
discovered line endings, or even expanded quoted strings.

A more promising approach would be to call a macro whenever
we increment line number that also increments the location
struct and re-calculates the column offset.